https://jira.aplana.com/browse/SBRFNDFL-74 Реализовать DAO интерфейс для работы с данными формы РНУ-НДФЛ, также добавил классы модели для хранения данных

+++ dao/src/main/java/com/aplana/sbrf/taxaccounting/dao/impl/NdflPersonDaoImpl.java
@@ -370,7 +370,7 @@ public class NdflPersonDaoImpl extends AbstractDao implements NdflPersonDao {
             personPrepayment.setRowNum(rs.getInt("row_num"));
             personPrepayment.setOperationId(SqlUtils.getLong(rs, "operation_id"));
 
-            personPrepayment.setSumm(rs.getBigDecimal("summ"));
+            personPrepayment.setSumm(rs.getLong("summ"));
             personPrepayment.setNotifNum(rs.getString("notif_num"));
             personPrepayment.setNotifDate(rs.getDate("notif_date"));
             personPrepayment.setNotifSource(rs.getString("notif_source"));
